refactor: export local mods view code in dedicated component
