+++ src-vue/src/views/ModsView.vue
@@ -50,24 +50,9 @@
 
         <!-- Mods content -->
         <div class="fc_mods__container">
-            <el-scrollbar v-if="show_local_mods">
-                <div>
-                    <p v-if="installedMods.length === 0">No mods were found.</p>
-                    <el-card v-else shadow="hover" v-for="mod in installedMods" v-bind:key="mod.name">
-                        <el-switch style="--el-switch-on-color: #13ce66; --el-switch-off-color: #8957e5" v-model="mod.enabled"
-                                   :before-change="() => updateWhichModsEnabled(mod)" :loading="global_load_indicator" />
-                        <el-popconfirm
-                            title="Are you sure to delete this mod?"
-                            @confirm="deleteMod(mod)"
-                        >
-                            <template #reference>
-                                <el-button type="danger">Delete</el-button>
-                            </template>
-                        </el-popconfirm>
-                        {{ mod.name }}
-                    </el-card>
-                </div>
-            </el-scrollbar>
+            <local-mods-view
+                v-if="show_local_mods"
+            />
 
             <thunderstore-mods-view 
                 v-else 
@@ -83,20 +68,19 @@
 
 <script lang="ts">
 import { defineComponent } from "vue";
-import { ElNotification } from "element-plus";
-import { invoke } from '@tauri-apps/api/tauri';
-import { GameInstall } from "../utils/GameInstall";
-import { NorthstarMod } from "../utils/NorthstarMod"
 import ThunderstoreModsView from "./ThunderstoreModsView.vue";
+import LocalModsView from "./mods/LocalModsView.vue";
 // @ts-ignore (why though?)
 import { SortOptions } from "../utils/SortOptions.d.ts";
 
 export default defineComponent({
     name: "ModsView",
-    components: { ThunderstoreModsView },
+    components: {
+        LocalModsView,
+        ThunderstoreModsView
+    },
     data() {
         return {
-            global_load_indicator: false,
             show_local_mods: true,
 
             // This is the model for the search input.
@@ -109,13 +93,9 @@ export default defineComponent({
         }
     },
     async mounted() {
-        this.$store.commit('loadInstalledMods');
         this.sortValue = this.sortValues[3].value;
     },
     computed: {
-        installedMods(): NorthstarMod[] {
-            return this.$store.state.installed_mods;
-        },
         sortValues(): {label: string, value: string}[] {
             return Object.keys(SortOptions).map((key: string) => ({
                 value: key,
@@ -124,66 +104,6 @@ export default defineComponent({
         }
     },
     methods: {
-        async updateWhichModsEnabled(mod: NorthstarMod) {
-            this.global_load_indicator = true;
-
-            // Setup up struct
-            let game_install = {
-                game_path: this.$store.state.game_path,
-                install_type: this.$store.state.install_type
-            } as GameInstall;
-
-            // enable/disable specific mod
-            try {
-                await invoke("set_mod_enabled_status", {
-                    gameInstall: game_install,
-                    modName: mod.name,
-                    // Need to set it to the opposite of current state,
-                    // as current state is only updated after command is run
-                    isEnabled: !mod.enabled,
-                })
-            }
-            catch (error) {
-                ElNotification({
-                    title: 'Error',
-                    message: `${error}`,
-                    type: 'error',
-                    position: 'bottom-right'
-                });
-                this.global_load_indicator = false;
-                return false;
-            }
-
-            this.global_load_indicator = false;
-            return true;
-        },
-        async deleteMod(mod: NorthstarMod) {
-            let game_install = {
-                game_path: this.$store.state.game_path,
-                install_type: this.$store.state.install_type
-            } as GameInstall;
-            await invoke("delete_northstar_mod", { gameInstall: game_install, nsmodName: mod.name })
-                .then((message) => {
-                    // Just a visual indicator that it worked
-                    ElNotification({
-                        title: `Success deleting ${mod.name}`,
-                        type: 'success',
-                        position: 'bottom-right'
-                    });
-                })
-                .catch((error) => {
-                    ElNotification({
-                        title: 'Error',
-                        message: error,
-                        type: 'error',
-                        position: 'bottom-right'
-                    });
-                })
-                .finally(() => {
-                    this.$store.commit('loadInstalledMods');
-                });
-        },
-
         /**
          * This method is called each time search input is modified, and
          * triggered filtered mods recomputing by updating the `searchValue`
